Scripted all NPCs in Cathedral w SD2
In order for the actions to take place during the Ashbringer event the
EventAI had to be ported to SD2 for all regular mobs. Added SD2 for
Fairbanks with the, hopefully, correct spells and abilities as seen on
retail. Values for kneeling and the random say w Ashbringer equipped
still needs more work (based on guestimates from the few videos that
still exist from Vanilla). Need to revisit all to clean up formatting.